+++ src/constants/data.tsx
@@ -168,22 +168,3 @@ export const MAIN_QCARD_QUESTION = [
         value: '네트워크 OSI 7계층에 대해 설명해 주세요.',
     },
 ];
-
-// TODO - categoryId 추가 요청
-export const QUESTION_CATEGORY_DETAIL = [
-    {
-        questionId: 21,
-        title: 'Spring DI에 대해 알려주세요',
-        category: '네트워크',
-    },
-    {
-        questionId: 22,
-        title: '네이버에 접속할때 일어나는 모든 과정에 대해 아는만큼 설명해주세요.',
-        category: '네트워크',
-    },
-    {
-        questionId: 23,
-        title: 'TCP 통신의 3-way Handshake 과정에 대해 설명해주세요.',
-        category: '네트워크',
-    },
-];
